+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 CHANGELOG:
 
+**Vesrion x.x.x**
+- 11/7/2024
+- Updated ClientList.js to also use DOB, email, and phone to pull the first and last name
+
 **Version x.x.x**
 - 11/7/2024
 - Updated ClientList.js to pull data from the database